+++ pkg/analyzer/test/src/summary/elements/metadata_test.dart
@@ -8612,7 +8612,7 @@ part 'b.dart';
 
     // The difference with the test above is that we ask the part first.
     // There was a bug that we were not loading library directives.
-    expect(library.parts[0].metadata, isEmpty);
+    expect(library.definingCompilationUnit.parts[0].metadata, isEmpty);
   }
 
   test_metadata_partOf_class() async {