+++ pkg/dev_compiler/lib/src/kernel/kernel_helpers.dart
@@ -20,6 +20,13 @@ Never throwUnsupportedAuxiliaryType(AuxiliaryType type) =>
 Constructor? unnamedConstructor(Class c) =>
     c.constructors.firstWhereOrNull((c) => c.name.text == '');
 
+/// Returns `true` when a class is eligible to be used as a mixin without
+/// having the mixin class modifier in Dart language versions < 3.0.
+extension ClassExtension on Class {
+  bool isLegacyMixinEligible(CoreTypes coreTypes) =>
+      !isMixinDeclaration && superclass == coreTypes.objectClass;
+}
+
 /// Returns the enclosing library for reference [node].
 Library getLibrary(NamedNode node) {
   for (TreeNode? n = node; n != null; n = n.parent) {