+++ pkg/linter/tool/migrate.dart
@@ -7,23 +7,17 @@
 import 'dart:io';
 
 import 'package:analyzer/error/error.dart';
+import 'package:args/args.dart';
 import 'package:collection/collection.dart';
 import 'package:linter/src/rules/analyzer_use_new_elements.dart';
 
 import 'checks/driver.dart';
 
 /// Lists files yet to be migrated to the new element model.
-main() async {
-  print('Unmigrated files:\n\n');
-  // (Start w/ rules.)
-  for (var rule in ruleFiles) {
-    if (!migratedFiles.contains(rule)) {
-      print(rule);
-    }
-  }
-
-  print('-' * 20);
-  print('-' * 20);
+main(List<String> args) async {
+  var parser = ArgParser()
+    ..addFlag('write',
+        abbr: 'w', help: 'Write updated `analyzer_use_new_elements.txt` file.');
 
   var errors = await getOldElementModelAccesses(rulesDir.absolute.path);
 
@@ -32,16 +26,30 @@ main() async {
     errorFiles.add(error.source.fullName);
   }
 
-  print('Migrated files:\n\n');
+  var migratedFiles =
+      ruleFiles.where((rule) => !errorFiles.any((f) => f.endsWith(rule)));
+
+  var options = parser.parse(args);
+  if (options['write'] == true) {
+    print("Writing to 'analyzer_use_new_elements.txt'...");
+    print('-' * 20);
+    File('analyzer_use_new_elements.txt')
+        .writeAsStringSync('${migratedFiles.join('\n')}\n');
+  } else {
+    print('Migrated files:\n');
+    print(migratedFiles.join('\n'));
+    print('-' * 20);
+    print('-' * 20);
+    print('\n');
+  }
+
+  print('Unmigrated files:\n\n');
   for (var rule in ruleFiles) {
-    if (!errorFiles.any((f) => f.endsWith(rule))) {
+    if (!migratedFiles.contains(rule)) {
       print(rule);
     }
   }
 }
-
-final List<String> migratedFiles =
-    File('analyzer_use_new_elements.txt').readAsLinesSync();
 
 final List<String> ruleFiles = rulesDir
     .listSync(recursive: true)