+++ pkg/analyzer/CHANGELOG.md
@@ -8,6 +8,10 @@
   use `CompilationUnitElement.libraryExports` instead.
 * Deprecated `LibraryElement.libraryImports`,
   use `CompilationUnitElement.libraryImports` instead.
+* Deprecated `LibraryElement.prefixes`,
+  use `CompilationUnitElement.libraryImportPrefixes` instead.
+* Deprecated `LibraryElement.parts`,
+  use `CompilationUnitElement.parts` instead.
 
 ## 6.9.0
 * `NormalFormalParameter` now implements `AnnotatedNode`.